Improved performance + Fixed memory leak.
Previously, all chunk->name variables would be allocated using calloc, now they're fixed-length as declared in the struct's declaration meaning that time shouldn't be wasted allocated and freeing that member now (in hindsight, this should've been done from the start as it is a fixed-length array).

A memory leak has also been fixed as I had forgotten to free the data member that is also allocated dynamically.