Fixed integer overflow (wraparound due to 2-4B)
An unsigned short was being used instead of an unsigned integer which restricted the max size of chunks being erased to USHORT_MAX (2^16-1) whereas the PNG specification uses 4 octets to specify chunk sizes.
Also, the strip_chunk function failed to account for chunk indexes being passed, this is also fixed.